feat: support multi room in the inspector and websocket
