+++ server/src/actors/mod.rs
@@ -5,4 +5,4 @@ pub use client_ws_actor::ClientWsActor;
 pub use game_actor::GameActor;
 
 pub mod room_manager_actor;
-pub use room_manager_actor::{CreateRoom, JoinRoom, RoomManagerActor};
+pub use room_manager_actor::{CreateRoom, JoinRoom, ListRooms, RoomManagerActor};
