+++ spectator/room/index.html
@@ -120,7 +120,7 @@
         disconnected
     </div>
 
-    <script src="index.js"></script>
+    <script src="./room/index.js"></script>
 </body>
 
 </html>
